random(4) fenestrasX: Push root seed version to arc4random(3)
Push the root seed version to userspace through the VDSO page (where
present*).

arc4random(3) obtains a pointer to the root seed version published by
the kernel in the shared page at allocation time.  Like arc4random(9),
it maintains its own per-process copy of the seed version corresponding
to the root seed version at the time it last rekeyed.  On read requests,
the process seed version is compared with the version published in the
shared page; if they do not match, arc4random(3) reseeds from the
kernel before providing generated output.

32-bit userspaces reads the 64-bit seed version in a multi-step process
to get a consistent snapshot of the version.  As the version changes
extremely infrequently, this should not have any problem with torn
results.  As the duplicate 32-bit read should be cached, this should not
be significantly slower than a plain 64-bit read.

This change does not implement the fenestrasX concept of PCPU userspace
generators seeded from a per-process base generator.  That change is
left for future discussion/work.

*N.B., not all FreeBSD architecture sysentvecs provide a VDSO / shared
page at this time.  I'm not sure if there is a technical reason for this
or just the general trend of MD code to diverge over time.  In this
version of the patch, arc4random(3) proceeds even when the shared page /
auxarg pointer is unavailable.  Of course, that means the mechanism is
non-functional on those platforms.

https://reviews.freebsd.org/D22839